for testing comment use case is done

+++ forum-api/src/Applications/use_case/comments/_test/DeleteCommentUseCase.test.js
@@ -16,8 +16,8 @@ describe('DeleteCommentUseCase', () => {
     const mockThreadRepository = new ThreadRepository();
     const mockCommentRepository = new CommentRepository();
     // Mocking
-    mockThreadRepository.verifyThreadAvailability = jest.fn().mockImplementation(() => Promise.resolve());
-    mockCommentRepository.verifyCommentOwner = jest.fn().mockImplementation(() => Promise.resolve());
+    mockThreadRepository.verifyThreadAccess = jest.fn().mockImplementation(() => Promise.resolve());
+    mockCommentRepository.verifyCommentAccess = jest.fn().mockImplementation(() => Promise.resolve());
     mockCommentRepository.deleteComment = jest.fn().mockImplementation(() => Promise.resolve());
 
     // Creating use case instance
@@ -30,8 +30,8 @@ describe('DeleteCommentUseCase', () => {
     await deleteCommentUseCase.execute(useCasePayload);
 
     // Assert
-    expect(mockThreadRepository.verifyThreadAvailability).toHaveBeenCalledWith(useCasePayload.threadId);
-    expect(mockCommentRepository.verifyCommentOwner).toHaveBeenCalledWith(useCasePayload.commentId, useCasePayload.owner);
+    expect(mockThreadRepository.verifyThreadAccess).toHaveBeenCalledWith(useCasePayload.threadId, useCasePayload.owner);
+    expect(mockCommentRepository.verifyCommentAccess).toHaveBeenCalledWith(useCasePayload.commentId, useCasePayload.owner);
     expect(mockCommentRepository.deleteComment).toHaveBeenCalledWith(useCasePayload.commentId);
   });
 
@@ -46,7 +46,7 @@ describe('DeleteCommentUseCase', () => {
     const mockThreadRepository = new ThreadRepository();
     const mockCommentRepository = new CommentRepository();
     // Mocking
-    mockThreadRepository.verifyThreadAvailability = jest.fn().mockImplementation(() => Promise.reject(new NotFoundError('THREAD_NOT_FOUND')));
+    mockThreadRepository.verifyThreadAccess = jest.fn().mockImplementation(() => Promise.reject(new NotFoundError('THREAD_NOT_FOUND')));
 
     // Creating use case instance
     const deleteCommentUseCase = new DeleteCommentUseCase({
@@ -69,8 +69,8 @@ describe('DeleteCommentUseCase', () => {
     const mockThreadRepository = new ThreadRepository();
     const mockCommentRepository = new CommentRepository();
     // Mocking
-    mockThreadRepository.verifyThreadAvailability = jest.fn().mockImplementation(() => Promise.resolve());
-    mockCommentRepository.verifyCommentOwner = jest.fn().mockImplementation(() => Promise.reject(new AuthorizationError('FORBIDDEN')));
+    mockThreadRepository.verifyThreadAccess = jest.fn().mockImplementation(() => Promise.resolve());
+    mockCommentRepository.verifyCommentAccess = jest.fn().mockImplementation(() => Promise.reject(new AuthorizationError('FORBIDDEN')));
 
     // Creating use case instance
     const deleteCommentUseCase = new DeleteCommentUseCase({
